Add continuous benchmarks
Add GitHub Actions workflow to run continuous benchmarks and publish to a tracking website.
