+++ LondonTravel.Site.sln
@@ -8,8 +8,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 		.gitattributes = .gitattributes
 		.gitignore = .gitignore
+		.markdownlint.json = .markdownlint.json
 		.sonarrc = .sonarrc
 		.vsconfig = .vsconfig
+		benchmark.ps1 = benchmark.ps1
 		build.ps1 = build.ps1
 		CODE_OF_CONDUCT.md = CODE_OF_CONDUCT.md
 		Directory.Build.props = Directory.Build.props
@@ -71,6 +73,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "workflows", "workflows", "{
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LondonTravel.Site.AppHost", "src\LondonTravel.Site.AppHost\LondonTravel.Site.AppHost.csproj", "{4C191691-65C5-4C8A-85AD-D8C55E6AD6C4}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "perf", "perf", "{E3A488C8-EE0F-4620-B545-FCFE34294C9D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LondonTravel.Site.Benchmarks", "perf\LondonTravel.Site.Benchmarks\LondonTravel.Site.Benchmarks.csproj", "{C50CE8BE-36CF-411C-9AD3-791850CE6BAD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -89,6 +95,10 @@ Global
 		{4C191691-65C5-4C8A-85AD-D8C55E6AD6C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4C191691-65C5-4C8A-85AD-D8C55E6AD6C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4C191691-65C5-4C8A-85AD-D8C55E6AD6C4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C50CE8BE-36CF-411C-9AD3-791850CE6BAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C50CE8BE-36CF-411C-9AD3-791850CE6BAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C50CE8BE-36CF-411C-9AD3-791850CE6BAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C50CE8BE-36CF-411C-9AD3-791850CE6BAD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -101,6 +111,7 @@ Global
 		{4AD9EC06-329C-47FB-8549-1A1EDE980130} = {DD7EA67A-4FE1-4D4F-825E-51FF842107AB}
 		{22736F26-BC95-4792-915C-B6E61FEB523A} = {4F53FEF1-66BC-45ED-B563-A619E9A193B7}
 		{4C191691-65C5-4C8A-85AD-D8C55E6AD6C4} = {0EED3EF2-37DE-4E6E-B766-3E641D163515}
+		{C50CE8BE-36CF-411C-9AD3-791850CE6BAD} = {E3A488C8-EE0F-4620-B545-FCFE34294C9D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {612E5D09-113B-4494-BEE6-0376224726FE}
