+++ tests/LondonTravel.Site.Tests/Integration/TestServerFixture.cs
@@ -19,7 +19,7 @@ namespace MartinCostello.LondonTravel.Site.Integration;
 /// <summary>
 /// A class representing a factory for creating instances of the application.
 /// </summary>
-public class TestServerFixture : WebApplicationFactory<ApplicationCookie>, ITestOutputHelperAccessor
+public class TestServerFixture : WebApplicationFactory<Program>, ITestOutputHelperAccessor
 {
     /// <summary>
     /// Initializes a new instance of the <see cref="TestServerFixture"/> class.
